+++ CHANGELOG.md
@@ -8,7 +8,7 @@ with the `YY.MINOR.MICRO` scheme.
 
 <img src="https://img.shields.io/badge/calver-YY.MINOR.MICRO-22bfda.svg">
 
-## [25.3.8] - Unreleased
+## [25.3.8] - 2025-10-23
 
 ### Fixes
 - Fixed an out-of-memory crash when using Response::encode_rmp with large